// types.ts
export enum ArtStyle {
  REALISTIC = 'Hyper-Realistic',
  CARTOON = 'Vibrant Cartoon',
  THREE_D_PIXAR = '3D Pixar Style',
  ANIME = 'Japanese Anime',
  VINTAGE_PHOTO = 'Vintage Sepia Photograph',
  CLAYMATION = 'Claymation Stop-motion',
  FANTASY_ART = 'Digital Fantasy Art',
  NEON_PUNK = 'Neon Punk',
}

export type RegenerationModel = 'gemini' | 'imagen';